Mail has some scrolling fixed in message view, a bug fixed in message send, and color added to the header of message view. Rules updated by jolz to give it more stack space.
abook is the simple cli utility to modify the contents of the
addressbook.xml data file, by making queries to the addressbook service.

addressbook is the actual address book service.  it is spawned by mail and
abook if it is not already in memory, and stays in memory as a daemon
until it is specifically killed.